feat: use body of git message as pr description
This commit adds the feature of including the body of the git commit message as
the description of the PR being created.

It also fixes the issue of the 'Create pull request' step failing whenever the
git commit message is more than just the subject. The reason why it used to fail
when that happens was because the environment GIT_MSG contains that full raw
body content, i.e., subject and body. If the body is non-empty, this means that
GIT_MSG would be more than a single line long. However, tea is expecting a
maximum of a single line as input for the title of the PR, creating an
unexpected error and failing right away.

Fixing this issue is easy, all we have done here is to fix the formatting of the
`git log` command from `%B%` (raw body) to `%s` (subject) for the newly defined
COMMIT_MSG environment that replaces GIT_MSG; and `%b%` (body) for the newly
defined COMMIG_DESC environment.